CLI interface to parser
