+++ rust/cmd_parser/Cargo.toml
@@ -5,6 +5,8 @@ version = "0.1.0"
 [dependencies]
 nom = "4.0"
 num = "0.1.40"
+rustyline = "1.0.0"
+libc = "0.2.42"
 
 [profile.release]
 opt-level = 1
